Handle unexpected whitespace in signature from Sphinx
When we take:

    testnamespace::MyClass::MyClass(const MyClass &)

And ask Sphinx to render it to text we get:

    testnamespace::MyClass::MyClass(const  MyClass \xa0&)

Where \xa0 is the 'non breaking whitespace' character. I'm not sure why
we get a double space after 'const' either.

This stuff doesn't happen in the main 'documentation' folder but does in
my separate test case. I have no idea how to explain that.